requirements.txt
streamlit==1.10.0
ibm-watson==5.3.1
ibm-cloud-sdk-core==3.16.0
pandas
requests
python-dotenv